feat: collect additional process attributes (#3605)
* refactor: move ProcessDetector to node; export noopDetector for Browser

* feat: collect additional process attributes

* refactor: always return a process resource

This commit introduces a slight change in behavior. Previously the
ProcessDectector would validate that it received attributes that it
expected, and returned an empty Resource if it didn't. Based on the
types, and the implementation of the ProcessDetector, there will
always be a valid ProcessResource. Collection of many of the attributes
can be done as a best effort without violating the specification.

* chore: add changelog entry

* feat: collect process.execArgv in process.command_args

* refactor: omit process.command_line in favor of process.command_args

* Move changelog entry to unreleased

---------

Co-authored-by: Daniel Dyla <dyladan@users.noreply.github.com>
Co-authored-by: Marc Pichler <marc.pichler@dynatrace.com>